+++ src/lib/utils/tokens.ts
@@ -27,7 +27,6 @@ import {
   WSTETH,
   icETHIndex,
 } from '@/constants/tokens'
-import { useProtectionContext } from '@/lib/providers/protection'
 
 const cbBTC = getTokenByChainAndSymbol(base.id, 'cbBTC')
 
@@ -128,24 +127,6 @@ export function isTokenBtcOnBase(
     isAddressEqual(outputToken, btc2x.address) ||
     isAddressEqual(outputToken, btc3x.address)
   )
-}
-
-export function useIsTokenPairTradable(
-  outputTokenSymbol: string,
-  chainId: number,
-): boolean {
-  const { isForbiddenAddress, isRestrictedCountry, isUsingVpn } =
-    useProtectionContext()
-  if (isForbiddenAddress) return false
-  if (!isRestrictedCountry && !isUsingVpn) return true
-  // When tokenlists is used everywhere, we can just pass these objects as function
-  // arguments instead of the token symbol
-  const outputToken = getTokenByChainAndSymbol(chainId, outputTokenSymbol)
-
-  const outputTokenIsDangerous =
-    outputToken?.tags.some((tag) => tag === 'dangerous') ?? false
-
-  return !outputTokenIsDangerous
 }
 
 export function digitsByAddress(address: string): number {
